extracted_project/rashdrivig/requirements.txt
flask
flask-socketio
eventlet
numpy
pandas
scikit-learn
joblib
RPi.GPIO
smbus2
requests
